Extract class from NKF
